Moved everything into same controller. Its now possible to remove the last value form the account
