+++ Panken/Views/Shared/_Layout.cshtml
@@ -30,7 +30,7 @@
             </div>
             <div class="navbar-collapse collapse">
                 <ul class="nav navbar-nav">
-                    <li><a asp-area="" asp-controller="Transaction" asp-action="DepositWithraw">Deposit/Withraw</a></li>
+                    <li><a asp-area="" asp-controller="Home" asp-action="DepositWithraw">Deposit/Withraw</a></li>
                     @*<li><a asp-area="" asp-controller="Home" asp-action="About">About</a></li>
                     <li><a asp-area="" asp-controller="Home" asp-action="Contact">Contact</a></li>*@
                 </ul>
